+++ tests/test_holdertrade_oos_backtest_wiring.py
@@ -59,7 +59,8 @@ class _StubEngine:
         self.kwargs = kwargs
         self.record_holdings_snapshot = False
 
-    def set_exposure_table(self, table, verbose=False, replenish=False, trim_tolerance=None):
+    def set_exposure_table(self, table, verbose=False, replenish=False, trim_tolerance=None,
+                         budget_discount_replenish=False):
         self.exposure_table = table
 
     def run(self, **kwargs):